ec: split the handle_fpu call in make_current int a save_fpu and a load_fpu call
In the later implementation, the stored and restored fpu-states are (most
likely) not in the samve pd and thus are not paged at the same time. Thus this
commit moves the load_fpu call behind pd::make_current to make sure we are in
the correct pd when loading the fpu-state.